backend/ml_service/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
scikit-learn==1.3.0
numpy==1.24.3
google-generativeai==0.1.0
pydantic==2.3.0
python-dotenv==1.0.0
